IMP - BUG - On timing of Rain Updating
To RainDropComponent, added function so updating segments is
now at a randomly-selected time.  The problem is, the rain is still updated
visually on the original timer in the RainContainer.  So now the rain
still moves every second as RainContainer specifies, it is just
that some drops now move farther down because the segments update at diff
speeds.